wrs/requirements.txt
panda3d>=1.10.7 # vital, visualization
numpy>=1.22.2 # vital, matrix computation
scipy>=1.10.1
tqdm>=4.65.0
opencv-python>=4.10.0.82 # optional, required by vision
opencv-contrib-python>=4.10.0.82 # optional, required by vision
scikit-learn>=0.23.2
grpcio>=1.56.2
grpcio-tools>=1.34.0
PyYAML>=5.3.1 # grpc formatting
Rtree>=0.9.7 # required by trimesh
open3d>=0.12.0 # required for cloud processing
shapely>=1.7.1 #
networkx>=2.5.1 # vital
pyserial>=3.4 # required by xarm controllers
pandas>=1.3.4 # required for neuro ik
matplotlib>=3.3 # 2D graph
Pillow>=10.0.1
python-dateutil>=2.8.2
lxml>=5.2.1
pyglet>=2.0.15
psutil>=5.9.4
sympy>=1.11.1
torch==2.0.1+cu117
torchviz>=0.0.2
decorator>=5.1.1
requests>=2.31.0
joblib>=1.2.0
ipython>=8.11.0
torchvision==0.15.2+cu117
linformer>=0.2.1
mujoco>=2.3.6